+++ src/main/java/org/opentmf/camunda/config/sso/plugin/SsoLogoutPluginConstants.java
@@ -6,4 +6,6 @@ package org.opentmf.camunda.config.sso.plugin;
 public class SsoLogoutPluginConstants {
 
   public static final String ID = "sso-logout-plugin";
+
+  private SsoLogoutPluginConstants() {}
 }